Update 0-att-bypass.sh
Test for UDM-P/SE should just confirm existence of eth8. Sometimes if the 802.1x authentication isn't delivered on time (like installing this script to fix a failed authentication) the ONT times out the link and it reads "Link detected: no". But eth8 still exists, so UDMPro is the right conf file to write.

On a UDM, testing eth8 will say "Cannot get link status" because there is no eth8. So looking for "Link detected" on eth8 is a good test, whether that status is yes or no.